+++ src/TitleDetails/ReceivedPiecesList/ReceivedPiecesList.test.js
@@ -28,6 +28,8 @@ describe('Given Received Pieces List', () => {
     const { getByText } = renderPiecesList();
 
     // header is rendered
+    expect(getByText('ui-receiving.piece.copyNumber')).toBeDefined();
+    expect(getByText('ui-receiving.piece.caption')).toBeDefined();
     expect(getByText('ui-receiving.piece.barcode')).toBeDefined();
     expect(getByText('ui-receiving.piece.enumeration')).toBeDefined();
     expect(getByText('ui-receiving.piece.format')).toBeDefined();