+++ src/TitleDetails/constants.js
@@ -27,6 +27,9 @@ export const ORDER_FORMAT_TO_PIECE_FORMAT = {
 };
 
 export const PIECE_COLUMNS = {
+  caption: 'caption',
+  chronology: 'chronology',
+  copyNumber: 'copyNumber',
   enumeration: 'enumeration',
   receiptDate: 'receiptDate',
   receivedDate: 'receivedDate',